+++ cotton-ssdp/src/lib.rs
@@ -49,7 +49,7 @@
 //!  - [ ] `Cow<'static>` for input strings?
 //!  - [ ] Hasher instead of `thread_rng`; hash over network interfaces sb unique
 //!  - [ ] Vary phase 1,2,3 timings but keep phase 0 timings on round numbers (needs _absolute_ wall time)
-//!  - [ ] Monotonic time instead of `Instant::now` (lifetime?)
+//!  - [x] Monotonic time instead of `Instant::now` (lifetime?) *Solved differently*
 //!  - [x] `smoltcp`/`no_std`, see <https://github.com/rust-lang/rust/pull/104265>
 //!  - [ ] IPv6, see UPnP DA appendix A
 //!
@@ -134,7 +134,6 @@ mod service;
 pub mod udp;
 
 /// Common code for triggering refreshes of [`Service`] and [`AsyncService`]
-#[cfg(feature = "std")]
 pub mod refresh_timer;
 
 #[cfg(feature = "async")]