+++ cross/rp2040-w5500/src/bin/rp2040-w5500macraw-dhcp-rtic.rs
@@ -55,7 +55,9 @@ mod app {
     #[init(local = [ storage: NetworkStorage = NetworkStorage::new() ])]
     fn init(c: init::Context) -> (Shared, Local, init::Monotonics) {
         defmt::println!("Pre-init");
-        let unique_id = unsafe { cross_rp2040_w5500::unique_flash_id() };
+        let unique_id = critical_section::with(|_| unsafe {
+            cross_rp2040_w5500::unique_flash_id()
+        });
         let mac = cotton_unique::mac_address(&unique_id, b"w5500-spi0");
         defmt::println!("MAC address: {:x}", mac);
 
@@ -135,8 +137,9 @@ mod app {
             hal::spi::FrameFormat::MotorolaSpi(embedded_hal::spi::MODE_0),
         );
 
-        let spi_device = 
-            embedded_hal_bus::spi::ExclusiveDevice::new_no_delay(spi_bus, spi_ncs);
+        let spi_device = embedded_hal_bus::spi::ExclusiveDevice::new_no_delay(
+            spi_bus, spi_ncs,
+        );
 
         let bus = w5500::bus::FourWire::new(spi_device);